Stateを切り出しました
